Some more Model code updates:
1. Made the shininess factor a float, since the code treats it as one and requires that anyway (colors will look better now)

2. Cleaned up some more short/int issues

git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@11532 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: 9b874295ae1f2724569422b2b1dcb1249b1e5243

+++ src/gui/rendering/ModelObject.hpp
@@ -25,7 +25,6 @@
 #include "Rmatrix.hpp"
 #include "Rvector3.hpp"
 
-// Since we are indexing into arrays with a short, is this realistic?
 #define MAX_VERTICES 200000  // Maximum vertices per object
 //#define MAX_VERTICES 80000  // Maximum vertices per object
 
@@ -83,7 +82,9 @@ typedef struct
    rgba_type mat_ambient;              // The ambient color of the material
    rgba_type mat_diffuse;              // The diffuse color of the material
    rgba_type mat_specular;             // The specular color of the material
-   unsigned short mat_shininess;       // The shininess factor of the material (between 0 and 100%, 100 being the highest)
+   // ??? This is a float throughout the code
+//   unsigned short mat_shininess;       // The shininess factor of the material (between 0 and 100%, 100 being the highest)
+   float mat_shininess;                // The shininess factor of the material (between 0 and 100%, 100 being the highest)
    int id_texture;                     // The id of the bound texture
 } material_type, *material_type_ptr;
